Reduce size of diff
